refactor: move TagList and searchTagList from util.ts to realm.ts
These are Realm-specific data and functions, not general utilities.
